pyproject.toml
[project]
name = "streamsets-databrickst-migration-agent"
version = "0.1.0"
description = "AI agent for migrating StreamSets Data Collector pipelines to Databricks"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "anthropic>=0.85.0",
    "fastapi>=0.115.0",
    "jinja2>=3.1.6",
    "portalocker>=3.2.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer[all]>=0.24.1",
    "uvicorn[standard]>=0.42.0",
]

[project.scripts]
migrate = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli", "agent", "catalog", "config", "models", "parser", "review", "state"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
]
